Вынос компонентов в отдельный папки и файлы. - по Offer
