+++ src/pages/main/main.tsx
@@ -1,9 +1,7 @@
-//import Card from '../../components/card/card.tsx';
 import ListOffer from '../../components/card/card.tsx';
 import {mockPreviewCard} from '../../mocks/mock-card.ts';
 import {directions} from '../../const.ts'
 import {Link } from 'react-router-dom';
-
 
 type CountArendaProps = {
   countArenda : number;
