+++ src/pages/offer/offer.tsx
@@ -1,77 +1,6 @@
-import { Link } from 'react-router-dom';
-import { useState } from 'react';
+import {Link } from 'react-router-dom';
+import OfferComment from '../../components/offer-comment/offer-comment';
 
-type CommentForOffer = {
-  id : string;
-  title: string;
-  rating: number;
-  description: string;
-}
-
-const nullCommentForOffer : CommentForOffer = {
-  id : '',
-  title : '',
-  rating : 0,
-  description : '',
-}
-
-function FormComment () :JSX.Element {
-  const [commentForOffer,setCommentForOffer] = useState(nullCommentForOffer);
-  return (
-    <form className="reviews__form form" action="#" method="post">
-                  <label className="reviews__label form__label" htmlFor="review">{commentForOffer.title}</label>
-                  <div className="reviews__rating-form form__rating">
-                    <input className="form__rating-input visually-hidden" name="rating" value="5" id="5-stars" type="radio" />
-                    <label htmlFor="5-stars" className="reviews__rating-label form__rating-label" title="perfect">
-                      <svg className="form__star-image" width="37" height="33">
-                        <use xlinkHref="#icon-star"></use>
-                      </svg>
-                    </label>
-
-                    <input className="form__rating-input visually-hidden" name="rating" value="4" id="4-stars" type="radio" />
-                    <label htmlFor="4-stars" className="reviews__rating-label form__rating-label" title="good">
-                      <svg className="form__star-image" width="37" height="33">
-                        <use xlinkHref="#icon-star"></use>
-                      </svg>
-                    </label>
-
-                    <input className="form__rating-input visually-hidden" name="rating" value="3" id="3-stars" type="radio" />
-                    <label htmlFor="3-stars" className="reviews__rating-label form__rating-label" title="not bad">
-                      <svg className="form__star-image" width="37" height="33">
-                        <use xlinkHref="#icon-star"></use>
-                      </svg>
-                    </label>
-
-                    <input className="form__rating-input visually-hidden" name="rating" value="2" id="2-stars" type="radio" />
-                    <label htmlFor="2-stars" className="reviews__rating-label form__rating-label" title="badly">
-                      <svg className="form__star-image" width="37" height="33">
-                        <use xlinkHref="#icon-star"></use>
-                      </svg>
-                    </label>
-
-                    <input className="form__rating-input visually-hidden" name="rating" value="1" id="1-star" type="radio" />
-                    <label htmlFor="1-star" className="reviews__rating-label form__rating-label" title="terribly">
-                      <svg className="form__star-image" width="37" height="33">
-                        <use xlinkHref="#icon-star"></use>
-                      </svg>
-                    </label>
-                  </div>
-                  <textarea className="reviews__textarea form__textarea" id="review" name="review" placeholder="Tell how was your stay, what you like and what can be improved">
-                    {commentForOffer.description}
-                  </textarea>
-                  <div className="reviews__button-wrapper">
-                    <p className="reviews__help">
-                      To submit review please make sure to set <span className="reviews__star">rating</span> and describe your stay with at least <b className="reviews__text-amount">50 characters</b>.
-                    </p>
-                    <button onClick={(evt) => {
-                      evt.preventDefault();
-
-                      setCommentForOffer(commentForOffer);
-                    }} className="reviews__submit form__submit button" type="submit" disabled = {false}>Submit</button>
-                  </div>
-                </form>
-  )
-}
 
 export default function Offer() : JSX.Element {
   return (
@@ -251,7 +180,7 @@ export default function Offer() : JSX.Element {
                   </li>
                 </ul>
 
-                <FormComment />
+                <OfferComment />
 
 
               </section>
